requirements.txt
absl-py==2.2.2
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
pycparser==2.22
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
zipp==3.21.0
